Create basic DOS 5 test scenario
I created a simple test scenario where a user joins the DOS 5 framework, confirms their company details, and then creates and deletes a user research service.

This is a simple proof of concept to show that I can create a Gatling scenario. There are a number of hard-coded things (e.g. the base URL) that I'll need to sort out later. The next step will be to create the other user flows we need (declaration questions, for example).

This works only on my local machine - I have a script to generate the users and suppliers for this test (PR to follow). Submitting it as a PR in the interests of getting fast feedback.