Setup for real load test
Increase the load to match the real load we want to use. Update the base URL to point at staging.

I ran this briefly for one user to confirm that this seems to be working OK.

Don't try to run this scenario against your local DMP, though. Not unless you want to exhaust your memory and force a reboot to recover :)